Changed Dirt to 2D array, easier access.
I changed the dirt container to a 2D array because thats what specs say
we should do and it made it easier to iterate through them. Also I
changed digger man to just a single pointer because the specs also said
to do that. Still trying to figure out how to set all appropriate dirt
invisible when running through the oilfield