dirt removal and boulder fall
Tried to implement dirt removal correctly so that the dirt is removed as
squares. Made the boulder fall and created random boulders in the map.
Didn't make the boulder disappear after ticks though. My laptop sucks
and moving the diggerman is super slow so i didn't thoroughly test if
the random boulders fell if you walked under them but they probably do.
Still have to make it so the diggerman cant walk through to boulder.