made the boulder fall after 30 ticks
the boulder falls after 30 ticks and goes away when it falls all the
way. Some boulders are being created in the mine shaft and outside of
the dirt and i didn't fix that yet. The boulders also shouldnt have any
dirt behind them and they should stop falling once they hit another
boulder. I didn't do any of that but i made them fall and go away lol.
The direction of the boulder is down in the spec so i changed that.