started on oil
Made the oil barrels show up on the map. Changed the implementation of
making boulders show up on map as well. Started on making barrels appear
when close, but the parameters were really big. I tried using the
isThere function for it but that didn't work because it's a little
different. Same for when the diggerman is close enough to grab.