Protestors can't walk through dirt
I hopefully correctly made the protestors not walk through the dirt.
Laptop is slow so protestors move slow so I tested as best I could!
Changed initial gold to 3 because specs. I split up the isThereProtestor
function into isDirtThere and isBoulderThere just because it was easier
for me to read it like that I hope that doesn't bother anyone.